created dataGeneration.js
